flaskserver/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors
SQLAlchemy==2.0.20
python-dotenv==1.0.1
Werkzeug==2.3.7
psycopg2-binary

